Resolved deletion of audio files issue
Solved issue of audio files getting deleted after moderation. The user now will have the choice to either delete or keep the audio file when it's downloaded from the URL. For audio files already present none will be deleted.

+++ IITD-speech-vone/utils.py
@@ -167,7 +167,7 @@ def get_transcript(bucket_name, bucket_folder, key_json, audio_folder):
 
 #***************************************************--------------------------------------------------------------------*********************************************************
 # to be used when the audio is locally stored
-def get_gender( input_audio):
+def get_gender(input_audio):
 	'''
 	This will return the predicted gender of the audio file
 	Basic Process:
@@ -183,8 +183,7 @@ def get_gender( input_audio):
 	input_audio_wav = input_audio_name + '.wav'
 	#remove_silence(input_audio_wav)
 	features = np.array([GC_exf.get_features(input_audio_name + '.wav')]).astype(float)
-	os.remove(input_audio_name + '.wav')
-	#os.remove(input_audio_name + '_rmsilence.wav')
+	#os.remove(input_audio_name + '.wav')
 	#model_path = pkg_resources.resource_filename('Indian_Speech_Lib', 'models/accept_reject')
 	model = joblib.load(open('models/Gender_classifier_rbf_model_python3.pk', 'rb'))
 	scaler = pickle.load(open('models/scaler_gender_rbf_python3.pk', 'rb'))
@@ -195,10 +194,11 @@ def get_gender( input_audio):
 		return 'female'
 
 # to be used when we are to download the audio from web url
-def get_gender_url(input_url):
+def get_gender_url(input_url, download_permanently = False):
 	get_audio_from_url(input_url,'temp.mp3')
 	gender = get_gender(input_audio = 'temp.mp3')
-	os.remove('temp.mp3')
+	if(download_permanently != True)
+		os.remove('temp.mp3')
 	return gender
 #***************************************************---------------------------------------------------------------------------*************************************************
 
@@ -260,7 +260,7 @@ def get_quality(input_audio):
 	input_audio_name, ext = os.path.splitext(input_audio)
 	input_audio_wav = input_audio_name + '.wav'
 	features = np.array(AR_exf.get_features(input_audio_name + '.wav')).astype(float)
-	os.remove(input_audio_name + '.wav')
+	#os.remove(input_audio_name + '.wav')
 	#model_path = pkg_resources.resource_filename('Indian_Speech_Lib', 'models/accept_reject')
 	model = joblib.load(open('models/AR_python3_SVM.pk', 'rb'))
 	scaler = pickle.load(open('models/AR_scaler.pk', 'rb'))
@@ -271,10 +271,11 @@ def get_quality(input_audio):
 		return 'accept'
 
 # to be used when we are to download the audio from web url
-def get_quality_url(input_url):
+def get_quality_url(input_url, download_permanently = False):
 	get_audio_from_url(input_url,'temp.mp3')
 	quality = get_quality(input_audio = 'temp.mp3')
-	os.remove('temp.mp3')
+	if(download_permanently != True) 
+		os.remove('temp.mp3')
 	return quality
 
 #*******************************************************************************
